+++ c_lib/masking/dom_operations.c
@@ -39,7 +39,7 @@ int dom_bool_and_##SHORT(                                                       
                                                                                 \
     const uint8_t order = mv_a->order;                                          \
     const uint8_t share_count = mv_a->share_count;                              \
-    const uint32_t share_bytes = mv_a->share_bytes;                             \
+    const uint16_t share_bytes = mv_a->share_bytes;                             \
     const uint32_t pair_count = (uint32_t)(share_count * order / 2);            \
     const uint32_t pair_bytes = pair_count * sizeof(TYPE);                      \
                                                                                 \
@@ -248,7 +248,7 @@ int dom_arith_mult_##SHORT(                                                     
                                                                                 \
     const uint8_t order = mv_a->order;                                          \
     const uint8_t share_count = mv_a->share_count;                              \
-    const uint32_t share_bytes = mv_a->share_bytes;                             \
+    const uint16_t share_bytes = mv_a->share_bytes;                             \
     const uint32_t pair_count = (uint32_t)(share_count * order / 2);            \
     const uint32_t pair_bytes = pair_count * sizeof(TYPE);                      \
                                                                                 \
